refactor: move headers to include/caches and update all references and docs
