+++ test/lfu_cache_tests.cpp
@@ -1,5 +1,5 @@
-#include "cache.hpp"
-#include "lfu_cache_policy.hpp"
+#include "caches/cache.hpp"
+#include "caches/lfu_cache_policy.hpp"
 
 #include <gtest/gtest.h>
 #ifdef CUSTOM_HASHMAP
