+++ test/nopolicy_cache_tests.cpp
@@ -1,4 +1,4 @@
-#include "cache.hpp"
+#include "caches/cache.hpp"
 
 #include <gtest/gtest.h>
 #include <stdexcept>
